requirements.txt
bcrypt
mysql-connector-python
screeninfo
tkcalendar
pandas